+++ Modules/Intent.Modules.Blazor/release-notes.md
@@ -1,3 +1,10 @@
+### Version 2.0.1
+
+- Improvement: Domain package reference automatically added to UI designer.
+- Improvement: Modelling context now added from the template, allowing for a consistent experience when not using auto-generated AI tasks.
+- Improvement: Having a space in the page name will no longer generate incorrect code.
+- Improvement: Component razor files are effectively written "Once off" (except for relevent directives)
+
 ### Version 2.0.0
 
 - Improvement: Component code-behind and styles are now emitted as separate `razor.cs` and `razor.css` files, allowing improved separation of concerns of C# logic and scoped CSS.